Remove api_requests
